// packages/fx-core/src/component/driver/teamsApp/interfaces/appdefinitions/IPublishingAppDefinition.ts
// Copyright (c) Microsoft Corporation.
// Licensed under the MIT license.

/**
 * Interface for Publishing Status in Teams App Catalog
 */
export interface IPublishingAppDenition {
  lastModifiedDateTime: Date | null;
  publishingState: PublishingState;
  /**
   * Teams app id in tenant App Catalog
   */
  teamsAppId: string;
  displayName: string;
}

export enum PublishingState {
  submitted = "submitted",
  published = "published",
  rejected = "rejected",
}